% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-exports.R
\docType{package}
\name{persistence-package}
\alias{persistence}
\alias{persistence-package}
\title{Persistence (deprecated)}
\description{
Given a non-oriented graph, calculates the optimal vertex partition using
persistence as the objective function.

\strong{This package is deprecated} and has been superseded by the
\pkg{scalednap} package, a strict superset that provides the same functions
and additional functionality. Please install \pkg{scalednap} instead
(\code{install.packages("scalednap")}); the same algorithm is also available
for Python (\code{pip install scalednap}).
}
\details{
See manual entries.
}
\author{
\strong{Maintainer}: Alessandro Avellone \email{alessandro.avellone@unimib.it}

Authors:
\itemize{
  \item Alessandro Avellone \email{alessandro.avellone@unimib.it}
  \item Paolo Bartesaghi \email{paolo.bartesaghi@unimi.it}
  \item Stefano Benati \email{stefano.benati@unitn.it}
  \item Rosanna Grassi \email{rosanna.grassi@unimib.it}
}

}
