% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistence-exports.R
\docType{package}
\name{persistence-package}
\alias{persistence}
\alias{persistence-package}
\title{Persistence}
\description{
Given a non-oriented graph, calculates the optimal vertex partition using the persistence as the objective function.
}
\details{
See manual entries.
}
\author{
\strong{Maintainer}: Alessandro Avellone \email{alessandro.avellone@unimib.it}

Authors:
\itemize{
  \item Paolo Bartesaghi \email{paolo.bartesaghi@unimi.it}
  \item Stefano Benati \email{stefano.benati@unitn.it}
  \item Rosanna Grassi \email{rosanna.grassi@unimib.it}
}

}
